backend/requirements.txt
flask
flask-cors
mysql-connector-python
reportlab
groq
python-dotenv